requirements.txt
# ── Phase 1-3: Data Collection & Preprocessing (local) ────────────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
rapidfuzz==3.6.1
imagehash==4.3.1
Pillow==10.2.0
tqdm==4.66.2
playwright>=1.44.0
playwright-stealth>=2.0.0

# ── Phase 4: API Baseline (local, no GPU) ─────────────────────────────────────
google-generativeai>=0.5.0    # Gemini API (free tier)

# ── Phase 4: Model Training (install on Google Colab, not locally) ────────────
# torch>=2.1.0
# torchvision>=0.16.0
# transformers>=4.38.0
# accelerate>=0.28.0
# peft>=0.10.0                # LoRA support

# ── Phase 5: Evaluation (local, CPU-only) ─────────────────────────────────────
nltk>=3.8.1
rouge-score>=0.1.2
# pycocoevalcap>=1.2          # optional: for CIDEr; install separately if needed
